Port to alpha 5 done
All should be working, I am beginning debugging now. This version was never even run
Only the red colour is used in LED control for now, because I'm using a simple OLA block instead of the final profile, which should be done today by Marsha.
The UNVTipcLamp profile I did yesterday was a good exercise, but it will be replaced by Marsha version

Signed-off-by: Luca Coppadoro <lcoppa@rocketmail.com>